pyproject.toml
[tool.poetry]
name = "PyDrocsid"
version = "2.3.0dev0"
description = "Python Discord Bot Framework based on pycord"
authors = ["Defelo <elodef42@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-only"
homepage = "https://github.com/PyDrocsid/library"
repository = "https://github.com/PyDrocsid/library"
packages = [{ include = "PyDrocsid" }]
include = ["PyDrocsid/emoji_map.json", "PyDrocsid/material_colors.yml", "PyDrocsid/translations/*"]

[tool.poetry.dependencies]
python = "^3.10"
# py-cord = {version = "^2.0.0b5", allow-prereleases = true}
py-cord = {git = "https://github.com/Defelo/pycord.git"}
SQLAlchemy = "^1.4.32"
PyYAML = "^6.0"
aiohttp = "^3.8.1"
sentry-sdk = "^1.5.8"
aenum = "^3.1.8"
aioredis = "^2.0.1"
asyncpg = "^0.25.0"
aiomysql = "^0.0.22"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
wemake-python-styleguide = "^0.16.1"
mypy = "^0.942"
SQLAlchemy = { extras = ["mypy"], version = "^1.4.32" }
types-PyYAML = "^6.0.5"

[tool.poe.tasks]
flake8 = "flake8 . --count --statistics --show-source"
isort = "isort ."
black = "black ."
format = ["isort", "black"]
mypy = "mypy ."
lint = ["format", "mypy", "flake8"]
pre-commit = ["lint"]

[tool.poe.tasks.setup]
shell = """
set -ex
poetry install --no-root
./pre-commit.sh install
"""
interpreter = "bash"

[tool.black]
target-version = ["py310"]
line-length = 120
skip-magic-trailing-comma = true

[tool.isort]
profile = "black"
py_version = 310
line_length = 120
lines_after_imports = 2
reverse_relative = true

[tool.mypy]
strict = true
ignore_missing_imports = true
no_incremental = true
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
